refactor(jmcore): host swap HTLC script primitive in jmcore
The swap HTLC script is a shared Bitcoin protocol primitive used by both the
taker (acquiring swap inputs) and the wallet (owning swap claim keys). It lived
in taker.swap.script, which forced jmwallet's swap-key test to import taker and
broke the jmwallet CI job (no taker installed there) at collection time.

Move SwapScript and the locktime bounds into jmcore.swap_script and re-export
from taker.swap.script / taker.swap.models for backwards compatibility, so no
component depends on taker for this primitive.
